pyproject.toml
[project]
name = "matta-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "kafka-python-ng>=2.2.3",
    "pyspark>=4.0.1",
]
